test(packaging): share native fixture helpers
